Refactoring app to be split into fragments
