+++ app/src/main/java/com/dillxn/tactilesynth/MenuFragment.java
@@ -1,0 +1,42 @@
+package com.dillxn.tactilesynth;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentActivity;
+import androidx.fragment.app.FragmentManager;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+
+/**
+ * A simple {@link Fragment} subclass.
+ * Use the {@link MenuFragment#newInstance} factory method to
+ * create an instance of this fragment.
+ */
+public class MenuFragment extends Fragment {
+    FragmentManager fragmentManager;
+    SynthFragment synthFrag;
+
+    public MenuFragment() {
+        // Required empty public constructor
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        fragmentManager = getParentFragmentManager();
+    }
+
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        View view = inflater.inflate(R.layout.fragment_menu, container, false);
+        synthFrag = (SynthFragment) fragmentManager.findFragmentByTag("synthPrime");
+
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_menu, container, false);
+    }
+}
